duplicate-word-detector_final/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
python-dotenv==1.0.0
PyMySQL==1.1.1

pythainlp==4.0.2

pandas==2.3.3
numpy==1.26.2

matplotlib==3.8.2
plotly==5.18.0
seaborn==0.13.2
wordcloud==1.9.4

pdfplumber==0.10.3
PyPDF2==3.0.1
pdfminer.six==20221105
pypdfium2==5.1.0

python-docx==1.1.0

pdf2image==1.16.3
pytesseract==0.3.10
Pillow==10.4.0

requests==2.32.5
psutil==5.9.8

blinker==1.9.0
click==8.3.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
Werkzeug==3.1.4

python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2

certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
fonttools==4.61.0
idna==3.11
kiwisolver==1.4.9
packaging==25.0
pycparser==2.23
pyparsing==3.2.5
six==1.17.0
typing_extensions==4.15.0
urllib3==2.6.0
